Rename `TypeOf` to be a part of `SystemFunctions`.
